Paths aren't just \w chars, but shebang lines must be 1 line
If the shebang line contains a character that matches \W (as the period did in my case), the grammar failed even though most characters that match \W are valid in paths.  It might be smart to make it aware of what characters are allowed in paths on the current operating system, but it doesn't really matter since it is being run by rperl not the command on the shebang anyway.